requirements.txt
streamlit
ultralytics
transformers
torch
pillow
opencv